Add the environment variable $PSM_RANKS_PER_CONTEXT.
Applied by hand from the following email.

From: Mark Debbage (QLogic)
Sent: Wednesday, December 21, 2011 10:45 AM
To: Adam T. Moody
Cc: Ira Weiny
Subject: RE: PSM contexts

Here's a patch to add the environment variable $PSM_RANKS_PER_CONTEXT.
You can use values in the range of 1-4 inclusive. I've tried this on a system
here and it looks good. It applies cleanly to the tip of the PSM mainline git tree
that you pull from.

We'll continue to track this in our internal bug 121444. I expect that I'll be
able to convince the team to add this patch to PSM, and then the patch
will subsequently be committed to our internal and external repos, and
available in our next IFS release.

Cheers and thanks for the suggestion! I've removed the mail list from the
cc as I can't send to that.

Mark.